# reseau_api/storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reseau Inventaire App API",
    "version": "1.0.0",
    "description": "Documentation statique minimale de l'API d'inventaire et de cartographie réseau."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8002",
      "description": "Serveur local Laravel"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token généré par Laravel Sanctum (Authorization: Bearer {token})"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "surname": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "role": { "type": "string" },
          "is_active": { "type": "boolean" }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Authentifier un utilisateur et récupérer un token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie ou informations invalides"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentification"],
        "summary": "Profil de l'utilisateur connecté",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Utilisateur connecté",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "description": "Non authentifié" }
        }
      }
    },
...
  }
}